+++ src/main/java/com/bsixel/mysticism/common/items/orbs/ForceOrb.java
@@ -1,7 +1,7 @@
 package com.bsixel.mysticism.common.items.orbs;
 
-import com.bsixel.mysticism.common.capability.mana.Force;
-import com.bsixel.mysticism.common.capability.mana.ManaCapability;
+import com.bsixel.mysticism.common.api.capability.mana.Force;
+import com.bsixel.mysticism.common.api.capability.mana.ManaCapability;
 import com.bsixel.mysticism.init.registries.ItemRegistry;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.util.ITooltipFlag;
@@ -9,10 +9,8 @@ import net.minecraft.client.util.InputMappings;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.ItemUseContext;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.util.ActionResult;
-import net.minecraft.util.ActionResultType;
 import net.minecraft.util.Hand;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.StringTextComponent;
@@ -25,7 +23,6 @@ import org.lwjgl.glfw.GLFW;
 
 import javax.annotation.Nullable;
 import java.util.List;
-import java.util.Random;
 
 public abstract class ForceOrb extends Item {
     Force force = Force.BALANCE;
@@ -69,8 +66,8 @@ public abstract class ForceOrb extends Item {
         ItemStack stack = player.getHeldItem(handUsed);
         if (!player.getCooldownTracker().hasCooldown(this) && stack.hasTag()) {
             player.getCapability(ManaCapability.mana_cap).ifPresent(playerMana -> {
-                float missingPlayerMana = playerMana.getMaxMana()-playerMana.getCurrentMana();
-                float manaToAdd = missingPlayerMana > 100 ? 100 : missingPlayerMana;
+                double missingPlayerMana = playerMana.getMaxMana()-playerMana.getCurrentMana();
+                double manaToAdd = missingPlayerMana > 100 ? 100 : missingPlayerMana;
                 CompoundNBT mystTag = stack.getChildTag("mysticism");
                 int currentOrbMana = mystTag.getInt("currentMana");
                 if (manaToAdd > 0) {
